requirements.txt
jira==2.0.0
requests==2.25.1
urllib3==1.26.5
python-dateutil==2.8.1
